moving with_lowercase_string_values to contedt_provider
